Bug Fixes for data-analytics workload to correct for changes to external distributions/repositories that prevented the initialization of the environment.
The updates addressed --
a) the renaming of an exception enum in the minio python client, and
b) the change in the default repository for helm.
c) corrected typos in documentation and comments in shell scripts

The updates do not affect the running of the benchmark workload, and consequently the benchmark results.